+++ src/shared/reducers/settings.js
@@ -1,4 +1,5 @@
 import uiReducer from './ui';
+import { ARCHIVES_REMOVE } from '../actions/types';
 
 function itemReducer(state = {}, action) {
   return {
@@ -13,6 +14,11 @@ export default function archivesReducer(state = {}, action) {
       ...state,
       [archiveId]: itemReducer(state[archiveId], action)
     };
+  }
+  if (action.type === ARCHIVES_REMOVE && state[action.payload]) {
+    const newState = {...state};
+    delete newState[action.payload];
+    return newState;
   }
   return state;
 }
